[atom] Restructure atom/diffusion to be model-major
The layout mirrored sglang's `multimodal_gen/runtime/`, which groups by
component type (dits/, vaes/, encoders/, schedulers/). That is right for a repo
hosting ~30 models that share components; with one model it only fragments.
Measured before this change:

* 65% of the subsystem (4,433 of 6,845 lines) was MiniMax-H3, spread across
  seven directories, so adding a second model would touch all seven;
* `stages/` contained **zero** PipelineStage subclasses -- all eight lived in
  `pipelines/minimax_h3.py`, and the 2,155 lines under `stages/minimax_h3/`
  were a pure-function library. The name was simply false;
* eight of sixteen directories held exactly one file, and `layers/` held none
  (a directory containing a directory containing one file);
* `models/loader.py` was fake-generic: 99 lines, one H3 function, importing
  from `models/dits/minimax_h3` -- a shared layer depending on a model.

Now the framework sits at the top level and each model is one package:

    atom/diffusion/
      config.py request.py pipeline.py attention.py ulysses.py mux.py
      engine/ entrypoints/ examples/
      models/minimax_h3/   arch, dit, vae, text_encoder, scheduler, loader,
                           pipeline (with its stages), and the H3 helpers

Sixteen directories become six, every single-file directory is gone, the stages
live with the pipeline that runs them, and the loader is honestly H3-specific.
This also matches ATOM's own convention -- `atom/models/` is one file per model
family, not a split by layer type.

The cost is losing path-parity with sglang, which made side-by-side diffing of
ported files easy. Judged not worth seven directories per model.

Pure relocation: no behaviour change, no signature change beyond
`stages/base.py` and `pipelines/base.py` merging into `pipeline.py` (they were
mutually dependent). 247 tests pass and the server's pipeline path resolves and
constructs.

A component graduates out of a model package into the shared layer when a
second model uses it, not in anticipation.
