+++ atom/diffusion/attention.py
