--- atom/diffusion/layers/attention/__init__.py
@@ -1,14 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (C) 2024-2025, Advanced Micro Devices, Inc. All rights reserved.
-
-from atom.diffusion.layers.attention.backend import (
-    AttentionBackend,
-    packed_varlen_attention,
-    resolve_attention_backend,
-)
-
-__all__ = [
-    "AttentionBackend",
-    "packed_varlen_attention",
-    "resolve_attention_backend",
-]
