--- atom/diffusion/models/dits/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (C) 2024-2025, Advanced Micro Devices, Inc. All rights reserved.
-
-"""Diffusion transformers (DiTs)."""
-
-from atom.diffusion.models.dits.minimax_h3 import MiniMaxH3DiTModel
-
-__all__ = ["MiniMaxH3DiTModel"]
