--- atom/diffusion/models/encoders/__init__.py
@@ -1,11 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (C) 2024-2025, Advanced Micro Devices, Inc. All rights reserved.
-
-"""Text / conditioning encoders for diffusion pipelines."""
-
-from atom.diffusion.models.encoders.minimax_h3_text import (
-    MINIMAX_H3_SELECTED_LM_LAYER,
-    MiniMaxH3TextEncoder,
-)
-
-__all__ = ["MINIMAX_H3_SELECTED_LM_LAYER", "MiniMaxH3TextEncoder"]
