+++ atom/diffusion/models/minimax_h3/arch.py
