+++ atom/diffusion/models/minimax_h3/presentation.py
