--- atom/diffusion/models/schedulers/__init__.py
@@ -1,16 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (C) 2024-2025, Advanced Micro Devices, Inc. All rights reserved.
-
-"""Diffusion samplers / schedulers."""
-
-from atom.diffusion.models.schedulers.euler_ancestral_h3 import (
-    MiniMaxH3EulerAncestralEta0Scheduler,
-    minimax_h3_euler_eta0_step,
-    minimax_h3_rf_v_to_x0,
-)
-
-__all__ = [
-    "MiniMaxH3EulerAncestralEta0Scheduler",
-    "minimax_h3_euler_eta0_step",
-    "minimax_h3_rf_v_to_x0",
-]
