--- atom/diffusion/models/vaes/__init__.py
@@ -1,20 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (C) 2024-2025, Advanced Micro Devices, Inc. All rights reserved.
-
-"""VAE components for diffusion pipelines."""
-
-from atom.diffusion.models.vaes.minimax_h3 import (
-    crop_to_canvas,
-    decode_audio_rows,
-    decode_video_rows,
-    denormalize_latents,
-    load_checkpoint_vae,
-)
-
-__all__ = [
-    "crop_to_canvas",
-    "decode_audio_rows",
-    "decode_video_rows",
-    "denormalize_latents",
-    "load_checkpoint_vae",
-]
