--- atom/diffusion/pipelines/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (C) 2024-2025, Advanced Micro Devices, Inc. All rights reserved.
-
-"""Composed diffusion pipelines."""
-
-from atom.diffusion.pipelines.base import ComposedPipeline
-
-__all__ = ["ComposedPipeline"]
