--- atom/diffusion/stages/__init__.py
@@ -1,12 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (C) 2024-2025, Advanced Micro Devices, Inc. All rights reserved.
-
-"""Composable pipeline stages for diffusion inference."""
-
-from atom.diffusion.stages.base import (
-    DiffusionBatch,
-    PipelineStage,
-    StageParallelism,
-)
-
-__all__ = ["DiffusionBatch", "PipelineStage", "StageParallelism"]
