+++ tests/test_diffusion_h3_fl2va.py
@@ -11,12 +11,12 @@ import numpy as np
 import pytest
 import torch
 
-from atom.diffusion.stages.minimax_h3.keyframe import (
+from atom.diffusion.models.minimax_h3.keyframe import (
     KEYFRAME_ENCODE_SEED,
     cover_crop_plan,
     scoped_encode_rng,
 )
-from atom.diffusion.stages.minimax_h3.packed_sequence import (
+from atom.diffusion.models.minimax_h3.packed_sequence import (
     FL2VA_KEYFRAME_SIGNATURES,
     FRAME_RESCALE,
     build_packed_sequence,
