+++ tests/test_diffusion_h3_ref2va_pipeline.py
@@ -16,19 +16,19 @@ import torch
 from torch import nn
 
 from atom.diffusion.config import ComponentConfig, DiffusionConfig
-from atom.diffusion.pipelines.minimax_h3 import (
+from atom.diffusion.models.minimax_h3.condition_noise import (
+    MINIMAX_H3_AUDIO_REF_COND_TIMESTEP,
+    MINIMAX_H3_IMGVID_COND_TIMESTEP,
+)
+from atom.diffusion.models.minimax_h3.pipeline import (
     ConditionEncodeStage,
     MiniMaxH3Pipeline,
     PackedSequenceStage,
     PlanStage,
     reference_materials,
 )
+from atom.diffusion.pipeline import DiffusionBatch
 from atom.diffusion.request import DiffusionJob
-from atom.diffusion.stages.base import DiffusionBatch
-from atom.diffusion.stages.minimax_h3.condition_noise import (
-    MINIMAX_H3_AUDIO_REF_COND_TIMESTEP,
-    MINIMAX_H3_IMGVID_COND_TIMESTEP,
-)
 from tests.test_diffusion_h3_pipeline import StubTextEncoder
 
 Image = pytest.importorskip("PIL.Image")
@@ -81,7 +81,7 @@ def write_reference_image(tmp_path):
 def build_ref2va(tmp_path, conditions, *, steps=3, duration=0.5):
     config = DiffusionConfig(
         model_path="<test>",
-        pipeline_class="atom.diffusion.pipelines.minimax_h3.MiniMaxH3Pipeline",
+        pipeline_class="atom.diffusion.models.minimax_h3.pipeline.MiniMaxH3Pipeline",
         components=[
             ComponentConfig(name="transformer", class_path="torch.nn.Identity")
         ],
@@ -176,7 +176,7 @@ def test_condition_encode_produces_blocks_and_both_row_kinds(tmp_path, monkeypat
         ],
     )
     monkeypatch.setattr(
-        "atom.diffusion.stages.minimax_h3.reference_encoding."
+        "atom.diffusion.models.minimax_h3.reference_encoding."
         "load_reference_waveform",
         lambda *a, **k: (torch.zeros(2, 32000), 32000),
     )
@@ -207,7 +207,7 @@ def test_visual_references_are_noise_augmented_and_audio_ones_are_not(
         ],
     )
     monkeypatch.setattr(
-        "atom.diffusion.stages.minimax_h3.reference_encoding."
+        "atom.diffusion.models.minimax_h3.reference_encoding."
         "load_reference_waveform",
         lambda *a, **k: (torch.zeros(2, 32000), 32000),
     )
@@ -232,7 +232,7 @@ def test_packed_layout_reserves_exactly_what_was_encoded(tmp_path, monkeypatch):
         ],
     )
     monkeypatch.setattr(
-        "atom.diffusion.stages.minimax_h3.reference_encoding."
+        "atom.diffusion.models.minimax_h3.reference_encoding."
         "load_reference_waveform",
         lambda *a, **k: (torch.zeros(2, 32000), 32000),
     )
@@ -286,8 +286,8 @@ def test_condition_without_a_uri_is_refused():
 def test_denoise_trims_the_reference_audio_off_the_prediction():
     """The DiT returns *all* audio rows but only the generated video rows, so
     the audio side needs an explicit trim the video side does not."""
-    from atom.diffusion.stages.minimax_h3.denoise import run_denoise_loop
-    from atom.diffusion.stages.minimax_h3.packed_sequence import (
+    from atom.diffusion.models.minimax_h3.denoise import run_denoise_loop
+    from atom.diffusion.models.minimax_h3.packed_sequence import (
         build_packed_sequence_ref2va,
     )
 
@@ -329,8 +329,8 @@ def test_denoise_trims_the_reference_audio_off_the_prediction():
 
 
 def test_denoise_rejects_missing_audio_reference_rows():
-    from atom.diffusion.stages.minimax_h3.denoise import run_denoise_loop
-    from atom.diffusion.stages.minimax_h3.packed_sequence import (
+    from atom.diffusion.models.minimax_h3.denoise import run_denoise_loop
+    from atom.diffusion.models.minimax_h3.packed_sequence import (
         build_packed_sequence_ref2va,
     )
 
@@ -359,7 +359,7 @@ def test_denoise_rejects_missing_audio_reference_rows():
 
 def test_video_reference_frame_sampling_feeds_both_consumers():
     """Qwen and the VAE must see the same decoded array, not two decodes."""
-    from atom.diffusion.stages.minimax_h3.reference_encoding import (
+    from atom.diffusion.models.minimax_h3.reference_encoding import (
         sample_reference_video_frames,
     )
